+++ seq/filter.go
@@ -2,7 +2,10 @@ package seq
 
 import "iter"
 
-func Filter[V any](seq Seq[V], predicate FilterFunc[V]) Seq[V] {
+// Filter is a function that filters elements of type V in the pull-type version of
+// an iter.Seq using a function V -> bool (the predicate). Elements for which the
+// predicate returns false are filtered out of the sequence.
+func Filter[V any, F ~func(V) bool](seq Seq[V], predicate F) Seq[V] {
 	return func() (V, bool) {
 		for {
 			value, valid := seq()
@@ -16,7 +19,8 @@ func Filter[V any](seq Seq[V], predicate FilterFunc[V]) Seq[V] {
 	}
 }
 
-func FilterSeq[V any](seq iter.Seq[V], predicate FilterFunc[V]) iter.Seq[V] {
+// FilterSeq is the Filter function designed to work on the push-style iter.Seq.
+func FilterSeq[V any, F ~func(V) bool](seq iter.Seq[V], predicate F) iter.Seq[V] {
 	return func(yield func(value V) bool) {
 		for v := range seq {
 			if predicate(v) {
@@ -28,7 +32,9 @@ func FilterSeq[V any](seq iter.Seq[V], predicate FilterFunc[V]) iter.Seq[V] {
 	}
 }
 
-func Filter2[K, V any](seq Seq2[K, V], predicate Filter2Func[K, V]) Seq2[K, V] {
+// Filter2 is the Filter function designed to work on the 2-elements version of the
+// pull-style version of iter.Seq.
+func Filter2[K, V any, F ~func(K, V) bool](seq Seq2[K, V], predicate F) Seq2[K, V] {
 	return func() (K, V, bool) {
 		for {
 			key, value, valid := seq()
@@ -42,7 +48,8 @@ func Filter2[K, V any](seq Seq2[K, V], predicate Filter2Func[K, V]) Seq2[K, V] {
 	}
 }
 
-func FilterSeq2[K, V any](seq iter.Seq2[K, V], predicate Filter2Func[K, V]) iter.Seq2[K, V] {
+// FilterSeq2 is the Filter2 function designed to work on the push-style iter.Seq.
+func FilterSeq2[K, V any, F ~func(K, V) bool](seq iter.Seq2[K, V], predicate F) iter.Seq2[K, V] {
 	return func(yield func(K, V) bool) {
 		for k, v := range seq {
 			if predicate(k, v) {