+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.2.2] - 2025-01-01
+- Improvement to type definition of `Modulator`, `filter`, `map` and `flatMap`.
+- `filter`, `map` and `flatMap` can now work with type aliases of the predicate and
+  mapping function.
+- Flatmap fixed to close holding channel and continue to return data after first 
+  mapping.
+- Lexer now adds the special `EOF` token at the end of the token stream to signal end
+  of input to downstream modulators and parser. This allows for the creation of 
+  downstream logic that needs to operate when all the tokens have been lexed; E.g.,
+  a modulator that reverses the stream of tokens.
+- `Ignore` Modulator for ignoring specific tokens in the token stream (such as whitespace).
+- `Reverse` example Modulator for reversing the token stream.
+
 ## [0.2.1] - 2024-12-30
 - Filtering, mapping and flat-mapping can now work on both the pull and push versions of iter.Seq
   and iter.Seq2.